Core/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
fal-client>=0.4.1
aiofiles>=23.2.1
pydantic>=2.9.0
sqlalchemy>=2.0.25
pillow>=10.0.0
httpx>=0.26.0
